* Updating database schema to use integers for status and type fields, per ticket #35.  This involves updating the installer, post(s), and comment(s) classes.
* Inserted a few class constants to map the integer values to static
names.  The format for the constant should be INDICATOR_VALUE, like:
   STATUS_DRAFT
   STATUS_PUBLISHED
   STATUS_PRIVATE
The invocation of the constant will be self-documenting as to what the
value applies to, so that two constants of the same name in different
classes will be disambiguated:
   Comment::SOME_CONSTANT
   Post::SOME_CONSTANT

* Fix a type in the User PHPDoc.

* Updating posts class to synthesize get() and do_query() into a single
function get().  Updated the post class to call get() rather than
do_query().



git-svn-id: http://svn.habariproject.org/habari/trunk@153 653ae4dd-d31e-0410-96ef-6bf7bf53c507