I made the DB class so you can call the class either statically or dynamically and it figures out what you want.
If you call it statically, it uses the internal $instance that is created when the index.php calls DB::create() with our config info.  So you can use DB:: to access the Habari core database from anywhere.  For example, DB::get_results() will always act on the Habari core database.

You can also create your own instance of DB to connect to a different database, but you have to manage the instance.  For example:

$mydb = new DB(...);
$foo = $mydb->get_results(...);
// Insert a value from the core db into the managed external db:
$foo->stuff = DB::get_value(...);

The habari_db class is now DB, and the filename has been changed as required for __autoload().

git-svn-id: http://svn.habariproject.org/habari/trunk@164 653ae4dd-d31e-0410-96ef-6bf7bf53c507