Importer fixes, tag fixes, post id insertion.
Post now gets the lastInsertId from the database and stores it in the object instance for use later.  This is useful for creating a post and then immediately adding tags to that post, since otherwise you wouldn't know the post id.  Comment should probably also do this so that the CommentInfo records can be properly attached, the idea being that descendants of QueryInfo need to manage themselves which internal field holds the autonumber index.

Tags are now associated with the post that they go to.  Remaining issues include insertion of duplicate tags -- tag_text is a unique index, and should be, but the methods used to ignore inserts of duplicate keys (the fastest way to insert a new tag_text) are different per DB engine.

Some minor URL class modifications to reduce code size and account for error conditions, but should be scrutinized.

git-svn-id: http://svn.habariproject.org/habari/trunk@443 653ae4dd-d31e-0410-96ef-6bf7bf53c507