Fixes for atom. The collection rule for Atom was using the {$action} special parameter, which was causing weird things to happen.  This needed to be changed in the default schema data.
The AtomHandler needed to be tweaked to work with the updated ActionHandler and URL systems.

Removed all the $option = Options::o() preparing for the event that we lose that method.

Threw in a change to the paginator and restored some default options to the Options class that were lost in the big merge, that may or may not be retained.

git-svn-id: http://svn.habariproject.org/habari/trunk@446 653ae4dd-d31e-0410-96ef-6bf7bf53c507