Switches around the output filter from a prefix to a postfix.  Instead of outputting $post->out_content we should now output $post->content_out.  This will sync better with the plugin filter names that affect those outputs, which in this case would be "post_content_out".
Themes will now load a theme.php file from the theme directory if one is present.  Setting the define() for THEME_CLASS will define the class that should be used to output the current theme.  This can be defined in the theme.php file included with the theme.  The K2 theme includes such a thing, which can be used as a sample for perusal (but we should convert at least one theme we ship to use the default so that we know it works and to demonstrate that you don't need a custom theme class to write a theme - you don't, by the way).

The default theme class now handles searching properly.  The K2 search.php template has been modified to accommodate the results.  Changed the DB schema to use the new rules (thanks to freakerz on issue 197).

Now using optional parameters in Themes instead of func_get_args(). 

Removed logic for retrieving any theme data to the Theme class instead of UserThemeHandler itself, which will allow custom theme classes to use the existing data gathering functions OR override those function with their own.  Native language overrides FTW!

git-svn-id: http://svn.habariproject.org/habari/trunk@491 653ae4dd-d31e-0410-96ef-6bf7bf53c507