Multi-site updates from issue 187.
Added the Site class to return info about the currently accessed site.  Several key functions added for retrieving the local directories and URL directories of parts of the system.

Added the option 'habari_url' which points to the URL of the Habari install (as opposed to the current site's URL, which may be different).  Added the option 'theme_dir' which is the URL to the current theme directory.

Made many changes in the admin so that URL::get() is being used to get the URLs for the admin pages based on their rewrite rules instead of outputting a base url and where we might hope the page would be located.

The Options class now extends Singleton.  All references to Options::o()-> have been updated not to do that any more, since this is not allowed in the Singleton architecture.

InstallHandler can now install databases for sub-sites by creating a site config directory and putting a config.php there. ( /user/sites/{sitename}/config.php )  The themes/plugins/classes directories for that sub-site are also loaded from that directory for the sub-site.  This is yet to be corrected to allow the sites to share those directories with the root install.  See issue 187 for details/discussion.

Put a check in Theme so that the installer doesn't display errors for not having any handler_vars to pass in.  It probably will eventually, but this should be accounted for anyway.



git-svn-id: http://svn.habariproject.org/habari/trunk@492 653ae4dd-d31e-0410-96ef-6bf7bf53c507