This is a big commit to centralize URL generation. The Options class should no longer be used to obtain or display URLs. The Site class is for "static" Habari URLs, and the URL class is for "dynamic" URLs (as controlled by the rewriterules table).
Site::get_url() returns a full URL
Site::get_path() returns a relative URL
Site::get_dir() returns a full filesystem directory

All three of the above are pluggable.  A plugin needs to register a 
function named "filter_site_url_NAME", where NAME is the parameter you 
would pass to the get_url() method.  Likewise for 
"filter_site_path_NAME" and "filter_site_dir_NAME".

Site::out_dir() outputs a full URL
Site::out_path() outputs a relative path of a URL
Site::out_dir() outputs a fill filesystem directory

Legal values are documented in the comments before each method.  The 
most common invocation is likely to be Site::get_url('habari'), which 
returns the full URL to the Habari installation.  This works for sites 
installed into the document root as well as sites installed into 
sub-directories. 

HTTPS is preserved if it was used to access the site, though there is no 
built-in provision to force a link to use HTTPS.  This is up to plugins.

Most -- if not all -- of the class files that need this functionality 
have been 
updated to use it.

This resolves issues #243 and #248.



git-svn-id: http://svn.habariproject.org/habari/trunk@541 653ae4dd-d31e-0410-96ef-6bf7bf53c507