Initial XMLRPC client.  Can be used to ping update services, among other XMLRPC functions.
Currently does not handle error responses, and only implements the client, not the server.

XMLRPCDate is a class used to hold a date.  This is used to differentiate a DateTime datatype when passing such a value via the client.

XMLRPCBinary is a class used to hold binary data.  This is used to differentiate a binary datatype when passing such a value via the client.  You can load this value form a file using: $xmlrpcbinary->load_from_file($filename);



git-svn-id: http://svn.habariproject.org/habari/trunk@599 653ae4dd-d31e-0410-96ef-6bf7bf53c507