Including the Apache license in the root of htdocs, so that folks only checking out that directory will obtain the license.  This should also ensure that any zipfile we produce includes our license file.
Adding a NOTICE file, per the Apache standard.  This lists non-Apache 
code we're using, and notes the licenses under which those bits are made 
available.  Currently only JQuery is differently licensed.  A copy of 
the MIT license under which JQuery is licensed is provided in the new 
/doc directory.

Also included in the NOTICE file is a list of committers, presented in 
chronological order for when they were added to the team.



git-svn-id: http://svn.habariproject.org/habari/trunk@610 653ae4dd-d31e-0410-96ef-6bf7bf53c507