The bulk of this commit adds support to disable comments on posts, per issue 230.  Thanks, Freakerz, for the patch supplied, which I tweaked.
Also fixed is a problem with r621 which placed $type into 
/system/admin/publish.php without initialization.

Inforecords were spitting errors if you tried to access an info value 
that was not assigned.  Now accessing an unassigned info value will 
return false.

Also included in this commit is a new static method User::commenter().  
It will return an array of values for use in comment forms.  See 
k2/comments.php for example usage.



git-svn-id: http://svn.habariproject.org/habari/trunk@622 653ae4dd-d31e-0410-96ef-6bf7bf53c507