Fixing up recent changes to theme directories and selection.
Chooses a fallback theme if the directory specified in options does not exist.

Stores the theme directory as a single directory rather than the full directory path.

Uses the theme directory instead of the theme name to construct the output for Site::get_url('theme') so that themes can have different names from the directory that contains them.

git-svn-id: http://svn.habariproject.org/habari/trunk@625 653ae4dd-d31e-0410-96ef-6bf7bf53c507