Added overloaded first parameter to URL::get() so that you can pass a string (same as prior convention) or an array of strings.  URL::get() iterates through the array, finds the first matching rule name that exists, and uses that to build the URL.  This lets you...
Added the ability for Post::permalink() to use a rewrite rule specific to its content_type.  So if a rule "display_entry" exists and the content_type of the post is "entry" then that rule is used instead of "display_posts_by_slug".  If "display_entry" doesn't exist, then it falls back to "display_posts_by_slug" as usual, which incidentally makes no sense as a name for this rule please change it kthxbye.

Added ability to pass lots of args data to URL::get() and (by extension) RewriteRule::build() and then only use the data required for the rule (omitting the querystring for the extra parameters).  See the new $useall parameter, which defaults to using the old build behavior.

Post::permalink now sends the entire post record plus individual date data to URL::get().  This allows you to change the permalink by altering the rule in the rewrite_rules table without changing any code.  Prior to this, permalinks in /{$year}/{$mon}/{$mday}/{$slug} format were not possible without code alterations.

Added Utils::getdate() to return same values as PHP's getdate(), but with 0-padded values 'mon0' and 'mday0'.  So if getdate($foo) returned '4' for the 'mon' value, then Utils::getdate($foo) would return '04' for the 'mon0' value.  Very useful for permalinks with fixed formatting.

Using the above changes, you can now create a rule specifically for entries that is /year/month/day/slug and a separate rule for pages that is simply /slug and have the system generate the appropriate URL to link to them based on their defined content-type.

Added but did not enable logic to select a rule based on the arguments passed to build it.  This would allow many rules to have the same name, but selection of the rule to build would be based on the best match of incoming arguments to the required arguments.  See RewriteRule::arg_match() where the functionality is disabled by a documented premature return, and URL::get().

Updated paginator on k2's tag template to use new pagecount code and rule names.

Added "just in case" typecasting to array in inforecords.

git-svn-id: http://svn.habariproject.org/habari/trunk@648 653ae4dd-d31e-0410-96ef-6bf7bf53c507