Adding new default Theme class with template failover capabilities.  Please see the thread here for more information:
http://groups.google.com/group/habari-dev/browse_thread/thread/3a6499e5f6facd4d

Notes:
Tag-based templates are not currently enabled in this commit.
SQLite schema is not yet updated.
Existing users that upgrade to this code may need to run the UPDATE script on the thread above before they see any significant change.  Custom themes are most likely to be affected.  IF YOU USE A CUSTOM THEME, you will want to watch the application of this commit carefully.
This commit includes additional MySQL rewrite rules for Atom.

git-svn-id: http://svn.habariproject.org/habari/trunk@651 653ae4dd-d31e-0410-96ef-6bf7bf53c507