BEFORE YOU `svn up` YOU MIGHT WANT TO READ...
Adds execution of dbdelta if the stored database version is less than the version specified in source.  Modified a few bits of the existing dbdelta() function to make it work properly with the newly organized database schema.  The new schema looks more like the output of SHOW CREATE TABLE than the prior incarnation.

If you change the database schema, be sure to change the db version in the Version class to reflect the current svn revision.

dbdelta only affects table structure.  It does not affect rewrite_rules.  If the INSERT statements in the schema change, you may need to apply them by hand.

Added freakerz's conditional regex rewrite rules.  They seem to be working well, and seem to have enabled pagination, or at the very least, enabled the pathway to pagination.

Added 404 template capability.  Add a 404.php to your theme directory and it will be shown when a request has nothing to serve.  Pretty handy.

This change is pretty extensive.  You will probably not need to drop your database to upgrade, but you may want to consider TRUNCATEing your rewrite_rules table and running the INSERTs at the end of the schema to recreate the data in that table.

SQLite is totally fubar'ed in this release because dbdelta() isn't finished for it and its schema hasn't been updated.  If you're running SQLite, you should NOT use update to this revision.

git-svn-id: http://svn.habariproject.org/habari/trunk@665 653ae4dd-d31e-0410-96ef-6bf7bf53c507