Added BigJibby's cron classes from issue 265, with some minor code convention changes, and an addition that allows you to specify either a callback function or a plugin filter hook name as the cron callback.
Added mysql schema changes for crontab table.  Added table name to database connections class.

Fixed an issue in DatabaseConnections that would return a previously set row class when an anonymous class was requested.

Changed plugin hook in Post to post_update_* for when a field of a post is updated.  Added post_inserted hook for after a post is inserted.

Plugin execution is currently synchronous with page load.  Setting the $async parameter of the CronTab::run_cron() that executes in index.php to true will make RemoteRequests to a cron URL.  This URL is not currently in the default rewrite rules set and the handler is not currently present.

git-svn-id: http://svn.habariproject.org/habari/trunk@673 653ae4dd-d31e-0410-96ef-6bf7bf53c507